third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/budget/get-cost-expected.txt

This is a testharness.js-based test.
PASS BudgetService.GetCost should return correct results. 
PASS BudgetService.GetCost should return Type Error if an invalid argument is provided. 
Harness: the test ran to completion.
